+++ src/geometry/ray.rs
@@ -3,7 +3,6 @@ use std::fmt::Debug;
 use ultraviolet::vec::Vec3;
 
 use crate::geometry::AngularExt;
-use crate::Float;
 
 #[derive(Copy, Clone, Debug, Default)]
 pub struct Ray {
@@ -21,21 +20,21 @@ impl Ray {
 
     #[inline]
     #[must_use]
-    pub fn at(&self, t: Float) -> Vec3 {
-        self.origin + self.direction * t as f32
+    pub fn at(&self, t: f32) -> Vec3 {
+        self.origin + self.direction * t
     }
 
     #[inline]
     #[must_use]
-    pub fn distance(&self, p: Vec3) -> Float {
-        (p - self.origin).mag() as Float
+    pub fn distance(&self, p: Vec3) -> f32 {
+        (p - self.origin).mag()
     }
 }
 
 impl AngularExt<Self> for Ray {
     #[inline]
     #[must_use]
-    fn angle_to(&self, other: Self) -> Float {
+    fn angle_to(&self, other: Self) -> f32 {
         self.direction.angle_to(other.direction)
     }
 }
